refactor: `getStats` tests
